+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 
 - [TD-4438] Included-in-parameters reference dataset validation
+- [TD-5509]
+  - Data structure link labels graphql query
+  - Data structure link creation using source and target structure IDs
 
 ### Changed
 
@@ -24,9 +27,6 @@
 - [TD-3541] Get data_structure_versions for `api/v2`
 - [TD-5599] Improve grant reindex memory usage and API for Task stats
 - [TD-5507] Grants API v2
-- [TD-5509]
-  - Data structure link labels graphql query
-  - Data structure link creation using source and target structure IDs
 
 ## [5.1.0] 2023-02-13