+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Fixed
 
+- [TD-5575] Filter implementations with execution permission on execution groups creation
 - [TD-5566] Remove template content from Remediation plan audit payload
 
 ## [5.2.1] 2023-03-08